requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7  # For PostgreSQL support
email-validator==2.0.0  # For email validation
APScheduler==3.10.1     # For periodic tasks
Flask-Session==0.5.0    # For server-side sessions 